Sync

+++ src/jsmn.h
@@ -2,6 +2,7 @@
 #define __JSMN_H_
 
 #include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -39,9 +40,9 @@ enum jsmnerr {
  */
 typedef struct {
 	jsmntype_t type;
-	int start;
-	int end;
-	int size;
+	int16_t start;
+	int16_t end;
+	int16_t size;
 #ifdef JSMN_PARENT_LINKS
 	int parent;
 #endif
@@ -52,9 +53,9 @@ typedef struct {
  * the string being parsed now and current position in that string
  */
 typedef struct {
-	unsigned int pos; /* offset in the JSON string */
-	unsigned int toknext; /* next token to allocate */
-	int toksuper; /* superior token node, e.g parent object or array */
+	int16_t pos; /* offset in the JSON string */
+	int16_t toknext; /* next token to allocate */
+	int16_t toksuper; /* superior token node, e.g parent object or array */
 } jsmn_parser;
 
 /**
